Clean up Edit message and fix role problem with subTo

+++ Commands/Moderation/editMessage.js
@@ -2,11 +2,20 @@ exports.run = (client, message,args) => {
   if(args.length <2 || args.length == 0 ){
     message.reply("It looks like there was a problem with your parameters. Please double check them.")
   }
+  var failToFindString = "Could not find message id in this channel. Please confirm you are sending command from correct channel.";
+  var findFailure = false;
   if(message.member.hasPermission('MANAGE_NICKNAMES'))
   {
     var id = args.shift();
-    message.channel.fetchMessage(id).then((msg)=>{
-      msg.edit(args.join(" ")).then(message.delete()).catch(console.error);
+
+    message.channel.fetchMessage(id).catch(function(error){
+      findFailure = true;
+      message.channel.send(failToFindString);
+    }).then((msg)=>{
+        if(findFailure){
+          return;
+        }
+        msg.edit(args.join(" ")).then(message.delete()).catch(console.error);
     });
     return;
   }
